requirements.txt
numpy 
opencv-python
Pillow
matplotlib
torch
torchvision
tqdm
sklearn
